CAL-54 Simplify the WKT Geo fields for streaming video metacard
The frame-center and location metacard attributes for MPEG-TS files/streams
can contain very large amounts of data. This provides a way to simplify the
data by setting a "distance tolerance" that groups near-by data points
together. The distance tolerance is configurable for both the MPEG-TS
input transformer and the MPEG-TS UDP stream monitor.