CAL-89 Prepare Alliance for 0.1 release
Alliance 0.1 will be released against DDF 2.9.1. Some of the services exported by the imaging-app can not be deployed because they were built against upcoming code in DDF 2.10.0. Comment out those services and create an integration test that affirms that all applications can be installed and started.

This additionally fixes an issue with the imaging-actionprovider-chip service.